+++ lib/src/utils/constants.dart
@@ -5,17 +5,21 @@ import 'package:meta/meta.dart';
 @internal
 const double kMenuScreenPadding = 8;
 
-/// Intrinsic width step.
+/// Intrinsic width step
 @internal
-const double kMenuWidthStep = 50;
+const double kMenuWidthStep = 250;
 
 /// Pull-down menu animation duration.
 @internal
-const Duration kMenuDuration = Duration(milliseconds: 250);
+const Duration kMenuDuration = Duration(milliseconds: 335);
 
-/// Pull-down menu animation curve.
+/// Pull-down menu animation curve (slightly modified [Curves.easeOutBack]).
 @internal
-const Curve kCurve = Curves.linearToEaseOut;
+const Curve kCurve = Cubic(0.175, 0.885, 0.265, 1.175);
+
+/// Pull-down menu close animation curve.
+@internal
+const Curve kCloseCurve = Curves.linearToEaseOut;
 
 /// Pull-down menu border radius.
 @internal