+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.1.0
+## 0.1.0-beta.2
 
-* Initial release.
+- **NEW**: Added shadow behind pull-down menu.
+- **IMPROVEMENT**: Update pull-down menu show/hide animations to be more similar to native version.
+- **IMPROVEMENT**: `PullDownMenuItem` and `SelectablePullDownMenuItem` - replace `InkWell` with `GestureDetector`.
+- **IMPROVEMENT**: Improve pull-down menu positioning on screen.
+- **FIX**: pull-down animating from the top instead of bottom for `PullDownMenuPosition.above`.
+
+## 0.1.0-beta.1
+
+- Initial release.